+++ riasec360/.env
@@ -4,7 +4,12 @@
 # Prisma supports the native connection string format for PostgreSQL, MySQL, SQLite, SQL Server, MongoDB and CockroachDB.
 # See the documentation for all the connection string options: https://pris.ly/d/connection-strings
 
-DATABASE_URL="mysql://root:Root_password1@localhost:3306/riasec?schema=public"
+DATABASE_URL="mysql://root:Root_password1@mysql:3306/riasec?schema=public"
+#DATABASE_URL="mysql://root:Root_password1@localhost:3306/riasec?schema=public"
 NEXTAUTH_SECRET="49raj2TxAIpE85AXzZDBGsWpOsn4+dDC3mGltHI7CfM="
 NEXTAUTH_URL="http:localhost:3000"
 #DATABASE_URL ="mysql://root:Root_password1@localhost:3306/riasec2?schema=public"
+MYSQL_ROOT_PASSWORD=Root_password1
+MYSQL_DATABASE=riasec
+MYSQL_USER=admin
+MYSQL_PASSWORD=admin
