+++ riasec360/package.json
@@ -39,6 +39,7 @@
     "lucide-react": "^0.376.0",
     "next": "^14.2.22",
     "next-auth": "^4.24.7",
+    "prisma": "^5.14.0",
     "react": "^18",
     "react-datepicker": "^7.0.0",
     "react-day-picker": "^8.10.1",
