+++ riasec360/src/app/gerenciarAplicacao/editarAplicacao/[id]/(usersTable)/columns.tsx
@@ -1,9 +1,9 @@
 "use client";
 
 import { ColumnDef } from "@tanstack/react-table";
-
 import { Checkbox } from "@/components/ui/checkbox";
 import { UsuarioResumido } from "@/app/types/types";
+import React from "react";
 
 export const columns: ColumnDef<UsuarioResumido>[] = [
   {
@@ -11,8 +11,11 @@ export const columns: ColumnDef<UsuarioResumido>[] = [
     header: ({ table }) => (
       <Checkbox
         checked={
-          table.getIsAllPageRowsSelected() ||
-          (table.getIsSomePageRowsSelected() && "indeterminate")
+          table.getIsAllPageRowsSelected()
+            ? true
+            : table.getIsSomePageRowsSelected()
+            ? "indeterminate"
+            : false
         }
         onCheckedChange={(value) => table.toggleAllPageRowsSelected(!!value)}
         aria-label="Select all"
