+++ riasec360/src/app/gerenciarTestes/adicionarCartoes/[id]/columns.tsx
@@ -3,8 +3,8 @@
 import { ColumnDef } from "@tanstack/react-table";
 
 import { Checkbox } from "@/components/ui/checkbox";
-import { Button } from "@/components/ui/button";
 import { Cartao } from "@/app/types/types";
+import React from "react";
 
 export const columns: ColumnDef<Cartao>[] = [
   {
@@ -12,8 +12,11 @@ export const columns: ColumnDef<Cartao>[] = [
     header: ({ table }) => (
       <Checkbox
         checked={
-          table.getIsAllPageRowsSelected() ||
-          (table.getIsSomePageRowsSelected() && "indeterminate")
+          table.getIsAllPageRowsSelected()
+            ? true
+            : table.getIsSomePageRowsSelected()
+            ? "indeterminate"
+            : false
         }
         onCheckedChange={(value) => table.toggleAllPageRowsSelected(!!value)}
         aria-label="Select all"
